requirements.txt
flask==3.1.0
python-dotenv==1.0.1
werkzeug==3.1.3
flask-Login==0.6.3
flask-SQLAlchemy==3.1.1
flask-wtf==1.2.2
psycopg2==2.9.10

